// frontend/src/app/api/models/class-schedule-request-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { RecurringScheduleTemplateDto } from '../models/recurring-schedule-template-dto';
export interface ClassScheduleRequestDto {
  endTime?: string;
  fitnessClassId?: string;
  startTime?: string;
  templateDto?: RecurringScheduleTemplateDto;
}