// frontend/src/app/api/model/fitnessClassesDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FitnessClassesDto { 
    name?: string;
    description?: string;
    durationMinutes?: number;
    capacity?: number;
    category?: string;
}